intial push for lesson 2
